requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
apscheduler==3.10.4
pytest==8.0.2
pytest-asyncio==0.23.5
